+++ src/Symfony/Bundle/FrameworkBundle/Translation/Translator.php
@@ -12,7 +12,6 @@
 namespace Symfony\Bundle\FrameworkBundle\Translation;
 
 use Psr\Container\ContainerInterface;
-use Symfony\Component\DependencyInjection\ContainerInterface as SymfonyContainerInterface;
 use Symfony\Component\HttpKernel\CacheWarmer\WarmableInterface;
 use Symfony\Component\Translation\Translator as BaseTranslator;
 use Symfony\Component\Translation\MessageSelector;
@@ -48,27 +47,16 @@ class Translator extends BaseTranslator implements WarmableInterface
      *   * debug:     Whether to enable debugging or not (false by default)
      *   * resource_files: List of translation resources available grouped by locale.
      *
-     * @param ContainerInterface $container A ContainerInterface instance
-     * @param MessageSelector    $selector  The message selector for pluralization
-     * @param array              $loaderIds An array of loader Ids
-     * @param array              $options   An array of options
+     * @param ContainerInterface $container     A ContainerInterface instance
+     * @param MessageSelector    $selector      The message selector for pluralization
+     * @param string             $defaultLocale
+     * @param array              $loaderIds     An array of loader Ids
+     * @param array              $options       An array of options
      *
      * @throws InvalidArgumentException
      */
-    public function __construct(ContainerInterface $container, MessageSelector $selector, $defaultLocale = null, array $loaderIds = array(), array $options = array())
+    public function __construct(ContainerInterface $container, MessageSelector $selector, string $defaultLocale, array $loaderIds = array(), array $options = array())
     {
-        // BC 3.x, to be removed in 4.0 along with the $defaultLocale default value
-        if (is_array($defaultLocale) || 3 > func_num_args()) {
-            if (!$container instanceof SymfonyContainerInterface) {
-                throw new \InvalidArgumentException('Missing third $defaultLocale argument.');
-            }
-
-            $options = $loaderIds;
-            $loaderIds = $defaultLocale;
-            $defaultLocale = $container->getParameter('kernel.default_locale');
-            @trigger_error(sprintf('Method %s() takes the default locale as 3rd argument since version 3.3. Not passing it is deprecated and will trigger an error in 4.0.', __METHOD__), E_USER_DEPRECATED);
-        }
-
         $this->container = $container;
         $this->loaderIds = $loaderIds;