+++ src/Symfony/Bundle/SecurityBundle/Tests/Security/FirewallContextTest.php
@@ -37,18 +37,6 @@ class FirewallContextTest extends TestCase
         $this->assertEquals($config, $context->getConfig());
     }
 
-    /**
-     * @expectedDeprecation Method Symfony\Bundle\SecurityBundle\Security\FirewallContext::getContext() is deprecated since version 3.3 and will be removed in 4.0. Use Symfony\Bundle\SecurityBundle\Security\FirewallContext::getListeners/getExceptionListener() instead.
-     * @group legacy
-     */
-    public function testGetContext()
-    {
-        $context = (new FirewallContext($listeners = array(), $exceptionListener = $this->getExceptionListenerMock(), new FirewallConfig('main', 'request_matcher', 'user_checker')))
-            ->getContext();
-
-        $this->assertEquals(array($listeners, $exceptionListener), $context);
-    }
-
     private function getExceptionListenerMock()
     {
         return $this